Remove python-Levenshtein and suppress warnings
This commit removes python-Levenshtein that improves performance of
fuzzywuzzy. However, this requires building some C code, which is
a bit of a pain on win32 systems. Given that fuzzywuzzy is only used
for matching FCC channels with locast channels and there isn't that
much data to be processed, a slower pure-python SequenceMatcher will be
used. Fuzzywuzzy normally warns about this, but we suppress this
warning, since there isn't much performance impact.