Rework UIDs
Instead of letting the user choose a UID, locast2dvr now generates
a UID as a uuid. This is set as the main uid and other tuners uids
are based on this main uid. The tuner uid is generated based on the
main uid and the tuner's dma.
In case a multiplexer is stated, the uid is always set to the main
uid.